finish

+++ src/app/api/todos/route.ts
@@ -2,8 +2,13 @@ import { NextRequest, NextResponse } from 'next/server';
 import { boolean, object, string } from 'yup';
 
 import prisma from '../../lib/prisma';
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 
 export async function GET(request: NextRequest) {
+  const user = await getUserSessionServer();
+
+  if (!user) return NextResponse.json('Unauthorized', { status: 401 });
+
   const searchParams = request.nextUrl.searchParams;
 
   const take = Number(searchParams.get('take') ?? '10');
@@ -22,6 +27,9 @@ export async function GET(request: NextRequest) {
     );
 
   const todos = await prisma.todo.findMany({
+    where: {
+      userId: user.id,
+    },
     take,
     skip,
   });
@@ -35,12 +43,18 @@ const postSchema = object({
 });
 
 export async function POST(request: NextRequest) {
+  const user = await getUserSessionServer();
+
+  if (!user) return NextResponse.json('Unauthorized', { status: 401 });
+
   try {
     const { description, done } = await postSchema.validate(
       await request.json()
     );
 
-    const todo = await prisma.todo.create({ data: { description, done } });
+    const todo = await prisma.todo.create({
+      data: { description, done, userId: user.id },
+    });
     return NextResponse.json(todo);
   } catch (error) {
     return NextResponse.json(error, { status: 400 });
@@ -48,9 +62,13 @@ export async function POST(request: NextRequest) {
 }
 
 export async function DELETE(request: NextRequest) {
+  const user = await getUserSessionServer();
+
+  if (!user) return NextResponse.json('Unauthorized', { status: 401 });
+
   try {
     const todo = await prisma.todo.deleteMany({
-      where: { done: true },
+      where: { userId: user.id, done: true },
     });
 
     return NextResponse.json(todo);
